Separate out recorder.
